death note system works as planned, only cosmetic and logic changes are needed
