+++ Content.Shared/DeathNote/DeathNoteComponent.cs
@@ -1,9 +1,11 @@
-namespace Content.Server.DeathNote;
+using Robust.Shared.GameStates;
+
+namespace Content.Shared.DeathNote;
 
 /// <summary>
 /// Paper with that component is Death Note.
 /// </summary>
-[RegisterComponent]
+[RegisterComponent, NetworkedComponent]
 public sealed partial class DeathNoteComponent : Component
 {
 }
